fix: remove unnecessary 'as u32' casts for AbiErrorCode
- AbiError.code is already AbiErrorCode, not u32
- PluginError.code changed from u32 to AbiErrorCode
- Remove all 'as u32' casts in Rust code and examples
- Rename _ABI_ERROR_SIZE to _ABI_STRUCT_SIZE (measures struct size)

The enum is #[repr(u32)] and FFI-safe, no casting needed.

+++ crates/polyplug/src/runtime.rs
@@ -1084,7 +1084,7 @@ mod tests {
 
     #[test]
     fn abi_ok_constant() {
-        assert_eq!(polyplug_abi::AbiErrorCode::Ok as u32, 0_u32);
+        assert_eq!(polyplug_abi::AbiErrorCode::Ok, 0_u32);
     }
 
     /// TH-06: Verify host callbacks in runtime.rs use HostInterface self-passing pattern.
